requirements.txt
pylint
pytest
black
nbval
pyspark
pandas
tabulate
pandas
databricks-sql-connector
python-dotenv